Move identifiable information to one file
